+++ src/cli/protocol/index.js
@@ -1,2 +1,5 @@
+exports.activate = require('./activate');
 exports.compile = require('./compile');
+exports.deactivate = require('./deactivate');
 exports.deploy = require('./deploy');
+exports.list = require('./list');