+++ src/cli/protocol/list/index.js
@@ -1,0 +1,13 @@
+// eslint-disable-next-line import/no-unresolved
+const { cli, action } = require('core/cli');
+const { execute, run } = require('./list');
+
+cli('protocol:ls', 'List transaction protocols', input => action(execute, run, input), {
+  requirements: {
+    forgeRelease: true,
+    runningNode: true,
+    rpcClient: true,
+    wallet: false,
+  },
+  options: [],
+});